% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ablandscape_control.R
\name{ablandscape.control}
\alias{ablandscape.control}
\title{Get defaults for antibody landscape fitting}
\usage{
ablandscape.control(
  max.titer.possible = 10,
  min.titer.possible = -4,
  max.slope = 1,
  error.sd = 1.1,
  confint.stepsize = 0.1,
  model.fn = NULL
)
}
\value{
Returns a list of default fitting options for making antibody
  landscapes.
}
\description{
This function returns all the options that can be set for fitting of antibody
landscapes and the default values for them.
}
