% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ablandscape_fit.R
\name{ablandscape.delta.fit}
\alias{ablandscape.delta.fit}
\title{Fitting an antibody delta landscape}
\usage{
ablandscape.delta.fit(
  titers1,
  titers2,
  coords,
  bandwidth,
  degree,
  error.sd,
  control = list()
)
}
\arguments{
\item{titers1}{A vector of titers against each antigen.}

\item{titers2}{A vector of titers against each antigen.}

\item{coords}{A matrix of the antigen coordinates of each antigen.}

\item{bandwidth}{The bandwidth of the local regression.}

\item{degree}{The degree of the polynomials to be used, normally 1 or 2.}

\item{control}{Control parameters: see (\code{\link{ablandscape.control}})}
}
\value{

}
\description{
Fit an antibody landscape that represents the difference in reactivity
between two serum samples on a log scale.
}
