% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ablandscape_fit.R
\name{ablandscape.fit}
\alias{ablandscape.fit}
\title{Fitting with antibody landscapes}
\usage{
ablandscape.fit(
  titers,
  coords,
  bandwidth,
  degree,
  error.sd,
  control = list(),
  acmap = NULL
)
}
\arguments{
\item{titers}{A vector of titers against each antigen.}

\item{coords}{A matrix of the antigen coordinates of each antigen.}

\item{bandwidth}{The bandwidth of the local regression.}

\item{degree}{The degree of the polynomials to be used, normally 1 or 2.}

\item{control}{Control parameters: see (\code{\link{ablandscape.control}})}
}
\value{

}
\description{
Fit an antibody landscape to titers and antigen positions.
}
