% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ablandscape.R
\name{ablandscape.fit_to_acmap}
\alias{ablandscape.fit_to_acmap}
\title{Fitting with antibody landscapes}
\usage{
ablandscape.fit_to_acmap(
  titers,
  acmap,
  bandwidth,
  degree,
  error.sd,
  control = list()
)
}
\arguments{
\item{titers}{A named vector of titers against each antigen.}

\item{acmap}{The antigenic map object containing the antigens titrated.}

\item{bandwidth}{The bandwidth of the local regression.}

\item{degree}{The degree of the polynomials to be used, normally 1 or 2.}

\item{control}{Control parameters: see (\code{\link{ablandscape.control}})}

\item{grid_spacing}{Grid spacing for surface interpolation}
}
\value{

}
\description{
Fit an antibody landscape to titers and antigen positions.
}
