% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ablandscape_par.R
\name{ablandscape.par}
\alias{ablandscape.par}
\title{Get defaults for antibody landscape plotting}
\usage{
ablandscape.par(cex = 1, cex.basemap = cex, cex.titer = cex * 2,
  asp.z = 1, zaxt = "linear", col.grid = "grey80",
  col.surface = "grey20", col.surface.grid = "grey20",
  opacity.basemap = 0.8, opacity.surface = 0.8,
  opacity.surface.grid = 0.6)
}
\value{
Returns a list of default plotting options for making antibody landscapes.
}
\description{
This function returns all the options that can be set for plotting of antibody
landscapes and the default values for them.
}
