\name{ablandscapes-package}
\alias{ablandscapes-package}
\alias{ablandscapes}
\docType{package}
\title{
\packageTitle{ablandscapes}
}
\description{
\packageDescription{ablandscapes}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ablandscapes}
\packageIndices{ablandscapes}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ablandscapes}

Maintainer: \packageMaintainer{ablandscapes}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
