% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titer_conversion.R
\name{as.logtiter}
\alias{as.logtiter}
\title{Convert raw titers to log titers}
\usage{
as.logtiter(titers)
}
\arguments{
\item{titers}{The raw titers to convert.}
}
\description{
Convert raw titers to log titers
}
