% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ablandscape.R
\name{build_lndscp}
\alias{build_lndscp}
\title{Build an antibody landscape}
\usage{
build_lndscp(object, grid_spacing = 1)
}
\arguments{
\item{object}{}

\item{grid_spacing}{}
}
\value{
Returns a list, with a matrix of x and y coordinates for each point
  to be plotted on the landscape, and a matrix of z coordinates, which are
  the landscape heights at each of these points.
}
\description{
This is a function for building and antibody landscape from a given set of antigenic coordinates and titres.
}
