% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_mean_titer_negll_without_sd}
\alias{calc_mean_titer_negll_without_sd}
\title{Calculate the total negative log-likelihood of a mean titer where the standard deviation 
is unknown. This is an internal function used by the function mean_titer.}
\usage{
calc_mean_titer_negll_without_sd(pars, max_titers, min_titers)
}
\description{
Calculate the total negative log-likelihood of a mean titer where the standard deviation 
is unknown. This is an internal function used by the function mean_titer.
}
