% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titer_lims.R
\name{calc_titer_diffs}
\alias{calc_titer_diffs}
\title{Get titer differences}
\usage{
calc_titer_diffs(titers1, titers2, control = list())
}
\arguments{
\item{titers1}{A numeric/character vector or matrix of pre titer measurements.}

\item{titers2}{A numeric/character vector or matrix of post titer measurements.}

\item{control}{A list of fitting options (see details).}
}
\value{
Returns a list of length two with values max_diffs and min_diffs, giving the numeric 
vectors of the upper and lower bounds of the difference in pre and post titers on the log scale.
}
\description{
Function for getting upper and lower limits of differences between measured titers on the log scale.
}
