% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_titer_loglik}
\alias{calc_titer_loglik}
\title{Calculate log likelihood of single fitted HI titer}
\usage{
calc_titer_loglik(max_titer, min_titer, pred_titer, error_sd)
}
\arguments{
\item{max_titer}{The upper bound of the measured titer.}

\item{min_titer}{The lower bound of the measured titer.}

\item{pred_titer}{The predicted titer.}

\item{error_sd}{The standard deviation of the error.}
}
\value{
Returns the log-likelihood of the measured titer given the measured titer 
bounds and error standard deviation supplied.
}
\description{
This is the base function for performing a maximum likelihood calculation for a 
single fitted HI titer given upper and lower limits of the measured value.
}
\details{
This function simply calculates to log likelihood of a predicted measurement 
given the upper and lower bounds of the measurement, the main assumption being that the 
associated error is normally distributed.
}
