% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_titer_set_negll}
\alias{calc_titer_set_negll}
\title{Calculate the total negative log-likelihood of a predicted titer set}
\usage{
calc_titer_set_negll(max_titers, min_titers, pred_titers, titer_weights,
  hi_error_sd)
}
\arguments{
\item{max_titers}{Numeric vector of the upper bounds of the measured titers.}

\item{min_titers}{Numeric vector of the lower bounds of the measured titers.}

\item{pred_titers}{Numeric vector of predicted titers.}

\item{titer_weights}{Numeric vector of titer weights.}

\item{hi_error_sd}{The standard deviation of the error.}
}
\description{
This is a base function to sum the total _negative_ log likelihood of a predicted 
and measured set of HI titers.
}
\details{
This function simply calculates to log likelihood of a series of predicted 
measurements given the upper and lower bounds of the measurements, the main assumption 
being that the associated error is normally distributed.
}
