% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ablandscape_confidence.R
\name{fit_ci2coords}
\alias{fit_ci2coords}
\title{Get landscape confidence interval}
\usage{
fit_ci2coords(coords, heights, negll, ag_coords, max_titers, min_titers,
  bandwidth, degree, crop2chull, control)
}
\description{
Function for calculating confidence intervals.
}
