% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ablandscape_confidence.R
\name{fit_height2point}
\alias{fit_height2point}
\title{Fit landscape to single coordinate point}
\usage{
fit_height2point(
  point_coords,
  lndscp_height,
  ag_coords,
  max_titers,
  min_titers,
  bandwidth,
  degree,
  error.sd,
  crop2chull = TRUE,
  control = list()
)
}
\description{
Fit landscape to single coordinate point
}
