% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lndscp2point.R
\name{fit_lndscp2coords}
\alias{fit_lndscp2coords}
\title{Fit landscape to a set of coordinates}
\usage{
fit_lndscp2coords(coords, ag_coords, max_titers, min_titers, bandwidth,
  degree, crop2chull = TRUE, control = list())
}
\description{
Fit landscape to a set of coordinates
}
