% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lndscp2point.R
\name{fit_lndscp2point}
\alias{fit_lndscp2point}
\title{Fit landscape to single coordinate point}
\usage{
fit_lndscp2point(point_coords, ag_coords, max_titers, min_titers,
  bandwidth, degree, crop2chull = TRUE, control = list())
}
\description{
Fit landscape to single coordinate point
}
