% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{h3coords2015}
\alias{h3coords2015}
\title{Antigenic coordinates of H3N2 influenza strains.}
\format{A matrix of antigenic coordinates}
\source{
\url{http://doi.org/10.1126/science.1256427}
}
\usage{
h3coords2015
}
\description{
A dataset containing the antigenic coordinates of influenza viruses.
}
\keyword{datasets}
