% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_testing.R
\name{lndscp.cv.test}
\alias{lndscp.cv.test}
\title{Perform cross-validation testing}
\usage{
lndscp.cv.test(
  titer_table,
  ag_coords,
  test.bandwidths,
  test.degrees,
  train.prop,
  num.runs,
  error.sd,
  control = list(),
  verbose = TRUE
)
}
\description{
Perform cross-validation testing
}
