% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lndscp_map.R
\name{lndscp3d_map}
\alias{lndscp3d_map}
\title{Generate a 3D map plot for an antibody landscape}
\usage{
lndscp3d_map(object, xlim, ylim, zlim, ...)
}
\arguments{
\item{object}{The antibody landscape object}

\item{...}{Graphical parameters}
}
\description{
Generate a 3D map plot for an antibody landscape
}
