% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lndscp_surface.R
\name{lndscp3d_surface}
\alias{lndscp3d_surface}
\title{Add a surface to a 3d landscape plot}
\usage{
lndscp3d_surface(
  data3js,
  object,
  crop2chull = TRUE,
  toggle = "Surface",
  grid_spacing = 1,
  padding = 0,
  options = list()
)
}
\arguments{
\item{data3js}{The r3js data object}

\item{object}{The antibody landscape object}

\item{...}{Additional plotting parameters}
}
\description{
Add a surface to a 3d landscape plot
}
