% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lndscp_titers.R
\name{lndscp3d_titers}
\alias{lndscp3d_titers}
\title{Add titer impulses to a 3d landscape plot}
\usage{
lndscp3d_titers(
  data3js,
  object,
  zlim,
  show.impulses = TRUE,
  toggle = "Titers",
  options = list()
)
}
\arguments{
\item{data3js}{The r3js data object}

\item{object}{The antibody landscape object}

\item{zlim}{z limits for the plot}

\item{...}{Additional plotting parameters}
}
\description{
Add titer impulses to a 3d landscape plot
}
