% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lndscp2point.R
\name{lndscp_checkChull}
\alias{lndscp_checkChull}
\title{Check if coordinates are within convex hull of ag coordinates}
\usage{
lndscp_checkChull(point_coords, ag_coords)
}
\description{
Check if coordinates are within convex hull of ag coordinates
}
