% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lndscp_contour.R
\name{lndscp_contour_ggplot}
\alias{lndscp_contour_ggplot}
\title{Plot an ablandscape contour plot}
\usage{
lndscp_contour_ggplot(
  fit,
  xlim = NULL,
  ylim = NULL,
  crop2chull = TRUE,
  padding = 1,
  grid_x = NULL,
  grid_y = NULL,
  grid_spacing = 0.5,
  show.grid = TRUE,
  grid.color = "grey90",
  grid.linewidth = 0.5,
  ...
)
}
\arguments{
\item{fit}{The antibody landscape fit object, created using [ablandscape.fit()]}

\item{xlim}{x limits for the plot}

\item{ylim}{y limits for the plot}

\item{crop2chull}{Crop the plot to the convex hull defined by the antigens titrated}

\item{padding}{Padding to apply to the fit}

\item{grid_x}{x grid points at which the ablandscapes fit should be calculated (passed to [predict_lndscp_grid()])}

\item{grid_y}{y grid points at which the ablandscapes fit should be calculated (passed to [predict_lndscp_grid()])}

\item{grid_spacing}{grid spacing for the ablandscapes fit (passed to [predict_lndscp_grid()])}

\item{show.grid}{Show grid lines behind contours}

\item{grid.color}{Grid line colors when shown}

\item{grid.linewidth}{Grid line width when shown}

\item{...}{Additional arguments passed to [ggplot2::geom_contour_filled()]}
}
\value{
Returns the ggplot object for the contour plot
}
\description{
Plot a contour plot from an antibody landscape fit using `ggplot2` and [ggplot2::geom_contour_filled()].
}
