% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lndscp.R
\name{lndscp_r3jsplot}
\alias{lndscp_r3jsplot}
\title{View method for the antibody landscape object}
\usage{
lndscp_r3jsplot(
  fit,
  xlim = NULL,
  ylim = NULL,
  zlim = NULL,
  padding = 1,
  crop2chull = TRUE,
  output = "widget",
  show.titers = TRUE,
  show.impulses = show.titers,
  show.surface = TRUE,
  show.map.antigens = TRUE,
  show.map.sera = TRUE,
  show.axis = TRUE,
  show.sidegrid = TRUE,
  aspect.z = 1,
  rotation = NULL,
  translation = NULL,
  zoom = NULL,
  toggles = TRUE,
  grid_spacing = 0.5,
  options = list()
)
}
\arguments{
\item{xlim}{The x axis limits}

\item{ylim}{The y axis limits}

\item{zlim}{The z axis limits}

\item{crop2chull}{Should the plot be cropped to the convex hull bounded by the points}

\item{object}{The antibody landscape object}

\item{...}{Additional plotting parameters}
}
\description{
View method for the antibody landscape object
}
