% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_titer.R
\name{mean.foldchange}
\alias{mean.foldchange}
\title{Function for calculating the mean foldchange from a group of titers}
\usage{
\method{mean}{foldchange}(titers1, titers2, titer_sd = NULL,
  control = list())
}
\arguments{
\item{titers1}{A character vector of starting titers}

\item{titers2}{A character vector of end titers}

\item{titer_sd}{The expected standard deviation of the fold-change responses around the mean, see details.}

\item{control}{Control parameters: see (\code{\link{ablandscape.control}})}
}
\value{
Returns the maximum likelihood mean fold-change response.
}
\description{
Function for calculating the mean foldchange from a group of titers
}
