% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_titer.R
\name{mean.titer}
\alias{mean.titer}
\title{Calculate the geometric mean titer}
\usage{
\method{mean}{titer}(titers, titer_sd = NULL, control = list())
}
\arguments{
\item{titers}{A character vector or matrix of HI titers.}

\item{titer_sd}{The expected standard deviation of the titers around the mean, see details.}

\item{control}{Control parameters: see (\code{\link{ablandscape.control}})}
}
\value{
Returns the maximum likelihood geometric mean titer.
}
\description{
This is the basic function for calculating the geometric mean titer of a 
set of titers based on maximum likelihood.
}
