% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_titer.R
\name{mean_fromlims}
\alias{mean_fromlims}
\title{Calculate maximum likelihood mean from measurement limits}
\usage{
mean_fromlims(upper_lims, lower_lims, measurement_sd = NULL, start_mean = NULL)
}
\arguments{
\item{upper_lims}{The upper measurement limits}

\item{lower_lims}{The lower measurement limits}

\item{measurement_sd}{The expected standard deviation of the measurement accounting for error and variation}
}
\value{
Returns the maximum likelihood mean.
}
\description{
This is the underlying function called by functions such as mean_titer and 
mean_foldchange.
}
