% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lndscp2point.R
\name{model_lndscp_height}
\alias{model_lndscp_height}
\title{Function to model the landscape height}
\usage{
model_lndscp_height(
  point_coords,
  ag_coords,
  titers,
  max_titers,
  min_titers,
  bandwidth,
  degree,
  error.sd,
  pars
)
}
\description{
Function to model the landscape height
}
