% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{negll_titer_lm}
\alias{negll_titer_lm}
\alias{negll_lndscp_height}
\title{Calculate the negative log likelihood of a linear model}
\usage{
negll_titer_lm(par, max_titers, min_titers, ag_coords, ag_weights, error_sd)

negll_lndscp_height(
  par,
  lndscp_height,
  max_titers,
  min_titers,
  ag_coords,
  ag_weights,
  error_sd
)
}
\arguments{
\item{par}{A vector of parameters, intercept followed by coefficients for each
coordinate dimension, or in the case of getting likelihood for a given height 
simply the coefficients for each coordinate dimension}

\item{max_titers}{Numeric vector of the upper bounds of the measured titers}

\item{min_titers}{Numeric vector of the lower bounds of the measured titers}

\item{ag_coords}{Matrix of antigenic coordinates relative to the landscape
coordinates being modelled}

\item{ag_weights}{A vector of weights to apply to each antigen, according to
their distance from the point being modelled}

\item{error_sd}{The expected standard deviation of titer error}
}
\description{
This is the base function used by the optimizer to calculate the negative
log likelihood of a given set of linear model parameters
}
