% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ablandscape_predict.R
\name{predict.ablandscape.fit}
\alias{predict.ablandscape.fit}
\title{Predict method for antibody landscapes}
\usage{
\method{predict}{ablandscape.fit}(
  object,
  coords,
  interval = "none",
  level = 0.95,
  negll.fit = FALSE,
  crop2chull = TRUE,
  control = list()
)
}
\arguments{
\item{object}{The antibody landscapes fit object}

\item{coords}{Antigenic coordinates for which to predict landscapes heights}

\item{interval}{Type of interval calculation, (one of "none" or "confidence")}
}
\value{
Produces a vector of predictions or a matrix of predictions and
  bounds with column names fit, lwr, and upr if interval is set.
}
\description{
Produce predicted values based on a antibody landscapes fit object.
}
