% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_testing.R
\name{summary.lndscp.cv}
\alias{summary.lndscp.cv}
\title{Print a summary of cv testing results}
\usage{
\method{summary}{lndscp.cv}(cv_results, crop2chull = TRUE, trim = 0)
}
\description{
Print a summary of cv testing results
}
