% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lndscp.R
\name{view.ablandscape}
\alias{view.ablandscape}
\title{View method for the antibody landscape object}
\usage{
\method{view}{ablandscape}(object, xlim, ylim, zlim, crop2chull = TRUE,
  ...)
}
\arguments{
\item{object}{The antibody landscape object}

\item{xlim}{The x axis limits}

\item{ylim}{The y axis limits}

\item{zlim}{The z axis limits}

\item{crop2chull}{Should the plot be cropped to the convex hull bounded by the points}

\item{...}{Additional plotting parameters}
}
\description{
View method for the antibody landscape object
}
