% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ablandscape_par.R
\name{ablandscape.par}
\alias{ablandscape.par}
\title{Get defaults for antibody landscape plotting}
\usage{
ablandscape.par(
  cex = 1,
  cex.basemap = cex,
  cex.basemap.ags = cex.basemap,
  cex.basemap.sr = cex.basemap,
  cex.titer = cex,
  cex.titer.impulse = cex,
  col.impulse = "grey20",
  lwd.impulse = 0.5,
  lwd.grid = 1,
  lwd.titer.outline = 1,
  asp.z = 1,
  zaxt = "linear",
  col.grid = "grey80",
  col.surface = "grey20",
  col.surface.grid = "grey20",
  shininess.surface = 30,
  sidegrid.lwd = 1,
  sidegrid.col = "grey95",
  sidegrid.at = NULL,
  opacity.basemap = 0.4,
  opacity.basemap.ags = opacity.basemap,
  opacity.basemap.sr = opacity.basemap,
  opacity.surface = 0.8,
  opacity.surface.grid = 0.6
)
}
\value{
Returns a list of default plotting options for making antibody landscapes.
}
\description{
This function returns all the options that can be set for plotting of antibody
landscapes and the default values for them.
}
